Module reql::error

ReQL Error Messages

These are the error messages returned by the driver. See their documentation for details.

Enums

ReqlAuthError
The client failed authentication with the server.

ReqlAvailabilityError
A server in the cluster is unavailable

ReqlCompileError
The query cannot be compiled by the server

ReqlDriverError
An error has occurred within the driver

ReqlError
The most generic error message in ReQL

ReqlInternalError
Query execution stopped due to an internal error, i.e., a server bug.

ReqlNonExistenceError
A ReqlQueryLogicError that results from accessing a non-existent field or something else that can be handled with the default command.

ReqlOpFailedError
The operation has failed due to cluster state, configuration or table availability.

ReqlOpIndeterminateError
The status of the operation cannot be verified due to cluster state, configuration or table availability.

ReqlPermissionsError
The user account does not have the permissions necessary to execute the query.

ReqlQueryLogicError
The query contains a logical impossibility, such as adding a number to a string.

ReqlResourceLimitError
Query execution caused a resource limit (for example, the array size limit) to be exceeded.

ReqlRuntimeError
The parent class of all runtime errors

ReqlTimeoutError
The query has timed out

ReqlUserError
An error produced by the error command.