Enum reql::error::ReqlRuntimeError

pub enum ReqlRuntimeError {
    QueryLogic(ReqlQueryLogicError),
    ResourceLimit(ReqlResourceLimitError),
    User(ReqlUserError),
    Internal(ReqlInternalError),
    Timeout(ReqlTimeoutError),
    Availability(ReqlAvailabilityError),
    Permissions(ReqlPermissionsError),
}

The parent class of all runtime errors

All errors on the server unrelated to compilation. Programs may use this to catch any runtime error, but the server will always return a more specific error class.

Variants

QueryLogic

ResourceLimit

User

Internal

Timeout

Availability

Permissions